+++ src/app/dashboard/page.tsx
@@ -1,33 +1,3 @@
-'use client';
-
-import Sidebar from '@/src/components/SideBar/SideBar';
-import TopNav from '@/src/components/TopNav/TopNav';
-import { css } from '@/styled-system/css';
-
-export default function dashboardPage() {
-  return (
-    <div className={containerCss}>
-      <Sidebar />
-      <div className={mainContainer}>
-        <TopNav navText="Dashboard" />
-        <div className={mainCss}></div>
-      </div>
-    </div>
-  );
+export default function DashboardPage() {
+  return <div>Đây là trang Dashboard chính</div>;
 }
-
-const containerCss = css({
-  display: 'flex',
-  height: '100vh',
-});
-
-const mainContainer = css({
-  display: 'flex',
-  flexDirection: 'column',
-  flex: 1,
-});
-
-const mainCss = css({
-  flex: 1,
-  backgroundColor: 'white.50',
-});
